+++ th2_data_services/provider/v5/commands/grpc.py
@@ -426,6 +426,8 @@ class GetMessagesGRPCObject(IGRPCProvider5Command, ProviderAdaptableCommand):
         search_direction: str = "NEXT",
         result_count_limit: int = None,
         keep_open: bool = False,
+        message_id: List[str] = None,
+        attached_events: bool = False,
         filters: List[Filter] = None,
     ):
         """GetMessagesGRPCObject constructor.
@@ -439,8 +441,9 @@ class GetMessagesGRPCObject(IGRPCProvider5Command, ProviderAdaptableCommand):
             result_count_limit: Result count limit.
             keep_open: If the search has reached the current moment.
                 It is need to wait further for the appearance of new data.
+            message_id: List of message ids to restore the search
+            attached_events: If true, it will additionally load attachedEventsIds.
             filters: Filters using in search for messages.
-
         """
         super().__init__()
         self._start_timestamp = start_timestamp
@@ -451,6 +454,8 @@ class GetMessagesGRPCObject(IGRPCProvider5Command, ProviderAdaptableCommand):
         self._result_count_limit = result_count_limit
         self._keep_open = keep_open
         self._filters = filters
+        self._message_id = message_id
+        self._attached_events = attached_events
 
     def handle(self, data_source: GRPCProvider5DataSource) -> List[MessageData]:
         api = data_source.source_api
@@ -467,6 +472,8 @@ class GetMessagesGRPCObject(IGRPCProvider5Command, ProviderAdaptableCommand):
             result_count_limit=self._result_count_limit,
             keep_open=self._keep_open,
             filters=self._filters,
+            message_id=self._message_id,
+            attached_events=self._attached_events,
         )
         for response in stream_response:
             if response.WhichOneof("data") == "message":
@@ -493,6 +500,8 @@ class GetMessages(IGRPCProvider5Command, ProviderAdaptableCommand):
         result_count_limit: int = None,
         keep_open: bool = False,
         filters: List[Filter] = None,
+        message_id: List[str] = None,
+        attached_events: bool = False,
         cache: bool = False,
     ):
         """GetMessages constructor.
@@ -507,8 +516,9 @@ class GetMessages(IGRPCProvider5Command, ProviderAdaptableCommand):
             keep_open: If the search has reached the current moment.
                 It is need to wait further for the appearance of new data.
             filters: Filters using in search for messages.
+            message_id: List of message ids to restore the search
+            attached_events: If true, it will additionally load attachedEventsIds.
             cache: If True, all requested data from rpt-data-provider will be saved to cache.
-
         """
         super().__init__()
         self._start_timestamp = start_timestamp
@@ -519,6 +529,8 @@ class GetMessages(IGRPCProvider5Command, ProviderAdaptableCommand):
         self._result_count_limit = result_count_limit
         self._keep_open = keep_open
         self._filters = filters
+        self._message_id = message_id
+        self._attached_events = attached_events
         self._cache = cache
 
         self._decoder = GRPCObjectToDictAdapter()
@@ -537,6 +549,8 @@ class GetMessages(IGRPCProvider5Command, ProviderAdaptableCommand):
             search_direction=self._search_direction,
             result_count_limit=self._result_count_limit,
             keep_open=self._keep_open,
+            message_id=self._message_id,
+            attached_events=self._attached_events,
             filters=self._filters,
         ).handle(data_source)
         for message in stream: